Implementing the Recipe Details, Shared Recipe card, Fix Login, Create Recipe, Comments, and Rating

+++ RecipesFinalProjectModels/IngredientLine.cs
@@ -10,7 +10,7 @@ namespace RecipesFinalProjectModels
     {
         public int Id { get; set; }
         public Ingredients Ingredient { get; set; }
-        public double Quantity { get; set; }
+        public  decimal Quantity { get; set; }
         public string Measure { get; set; } = string.Empty;
         public Recipes Recipe { get; set; }
